move tom-select setup into a stimulus controller
