+++ app/javascript/controllers/index.ts
@@ -2,4 +2,7 @@
 // Run that command whenever you add a new controller or create them with
 // ./bin/rails generate stimulus controllerName
 
-import { application } from "./application"
+import { application } from './application'
+
+import TagInputController from "./tag_input_controller"
+application.register("tag-input", TagInputController)
