+++ app/javascript/src/model_edit.ts
@@ -1,22 +1,4 @@
-import TomSelect from 'tom-select'
-import type { TomInput } from 'tom-select/dist/cjs/types'
-
 document.addEventListener('ManyfoldReady', () => {
-  document.querySelectorAll('select[data-tom-select]').forEach(
-    (element: TomInput) => {
-      if (element.tomselect != null) { return }
-      new TomSelect(element, { // eslint-disable-line no-new
-        addPrecedence: true,
-        create: true,
-        plugins: ['remove_button'],
-        selectOnTab: true,
-        onItemAdd: function () {
-          this.setTextboxValue('')
-          this.refreshOptions()
-        }
-      })
-    }
-  )
   // Editable names (and other fields):
   $('[contenteditable=true]').focus(function () {
     $(this).data('initialText', $(this).text().trim())
