+++ app/views/models/_tags_edit.html.erb
@@ -8,7 +8,7 @@
             multiple: true,
             class: "form-select",
             name: name ? "#{name}[]" : nil,
-            data: {"tom-select": "true"}
+            data: {controller: "tag-input"}
           } %>
   </div>
 </div>
